Move comment commands to its own extensions
Stll PoC comment (hard-coded to C-like line comments).
