+++ comment/package.json
@@ -1,0 +1,5 @@
+{
+  "main": "dist/index.js",
+  "types": "src/index",
+  "type": "module"
+}
