+++ demo/demo.ts
@@ -11,6 +11,7 @@ import {specialChars} from "@codemirror/next/special-chars"
 import {multipleSelections} from "@codemirror/next/multiple-selections"
 import {search, defaultSearchKeymap} from "@codemirror/next/search"
 import {autocomplete, startCompletion} from "@codemirror/next/autocomplete"
+import {toggleCommentCmd, commentCmd, uncommentCmd} from "@codemirror/next/comment"
 
 import {html} from "@codemirror/next/lang-html"
 import {defaultHighlighter} from "@codemirror/next/highlight"
@@ -54,7 +55,10 @@ let state = EditorState.create({doc: `<script>
     "Mod-Alt-[": foldCode,
     "Mod-Alt-]": unfoldCode,
     "Mod-Space": startCompletion,
-    "Shift-Mod-m": openLintPanel
+    "Shift-Mod-m": openLintPanel,
+    "Mod-/": toggleCommentCmd,
+    "Mod-Alt-/": commentCmd,
+    "Mod-Alt-Shift-/": uncommentCmd
   }),
   keymap(baseKeymap),
 ]})
@@ -62,56 +66,14 @@ let state = EditorState.create({doc: `<script>
 let view = (window as any).view = new EditorView({state})
 document.querySelector("#editor")!.appendChild(view.dom)
 
-
-// let tr = state.t().replace(6, 11, "from transaction")
-// console.log(tr.doc.toString()) // "hello editor"
-// let newState = transaction.apply()
-// view.dispatch(tr)
-
-enum CommentOption {
-  Toggle,
-  OnlyComment,
-  OnlyUncomment,
-}
-
-const toggleComment = function(option: CommentOption = CommentOption.Toggle) {
-  const lineCommentToken = "//"
-  let s = view.state
-  let f = s.selection.primary.from
-  let t = s.selection.primary.to
-  let l = s.doc.lineAt(f)
-  console.log(f, t, l.start)
-  let str = (l.content as string)
-  if (str.startsWith(lineCommentToken)) {
-    if (option != CommentOption.OnlyComment) {
-      let tr = view.state.t().replace(l.start, l.start + lineCommentToken.length, "")
-      view.dispatch(tr)
-    }
-
-  } else {
-    if (option != CommentOption.OnlyUncomment) {
-      let tr = view.state.t().replace(l.start, l.start, lineCommentToken)
-      view.dispatch(tr)
-    }
-  }
-}
-
-const comment = function() {
-  toggleComment(CommentOption.OnlyComment)
-}
-
-const uncomment = function() {
-  toggleComment(CommentOption.OnlyUncomment)
-}
-
 document.querySelector("#toggleComment")!.addEventListener("click", function() {
-  toggleComment()
+  toggleCommentCmd(view)
 });
 
 document.querySelector("#comment")!.addEventListener("click", function() {
-  comment()
+  commentCmd(view)
 });
 
 document.querySelector("#uncomment")!.addEventListener("click", function() {
-  uncomment()
+  uncommentCmd(view)
 });
