+++ package.json
@@ -49,7 +49,8 @@
     "./autocomplete": "./autocomplete/dist/index.js",
     "./lang-javascript": "./lang-javascript/dist/index.js",
     "./lang-css": "./lang-css/dist/index.js",
-    "./lang-html": "./lang-html/dist/index.js"
+    "./lang-html": "./lang-html/dist/index.js",
+    "./comment": "./comment/dist/index.js"
   },
   "license": "(MIT OR GPL-3.0)",
   "dependencies": {
